\name{nunique}
\alias{nunique}
\title{How Many Unique Values?}
\description{
  Returns the number of unique values in a vector.
}
\usage{
nunique(x)
}
\arguments{
  \item{x}{A vector.}
}
\value{
  An integer equal to \code{length(unique(x))}.
}
\author{Benjamin Rich <mail@benjaminrich.net>}

\keyword{ utilities }

% vim: tw=70 sw=2 ts=2 et :
